Rename query builder contract
